+++ crates/native/src/media/mod.rs
@@ -23,6 +23,8 @@ pub use self::{
         AudioTrack, AudioTrackId, Track, TrackOrigin, VideoTrack, VideoTrackId,
     },
 };
+#[cfg(target_os = "windows")]
+use crate::media::source::SYSTEM_AUDIO_DEVICE_ID;
 use crate::{
     Webrtc, api, api::NoiseSuppressionLevel, devices,
     frb_generated::StreamSink, pc::PeerConnectionId,
@@ -285,42 +287,68 @@ impl Webrtc {
         &mut self,
         caps: &api::AudioConstraints,
     ) -> anyhow::Result<Arc<AudioSource>> {
-        let device_id = if let Some(device_id) = caps.device_id.clone() {
-            device_id.into()
-        } else {
-            // `AudioDeviceModule` is not capturing anything at the moment, so
-            // the first available device (with `0` index) will be used.
-            if self.audio_device_module.recording_devices() < 1 {
-                bail!("Cannot find any available audio input device");
+        let processing = sys::AudioProcessing::new((&caps.processing).into())?;
+
+        let src = if caps.is_display {
+            #[cfg(not(target_os = "windows"))]
+            bail!("Display audio tracks aren't supported on your platform.");
+
+            #[cfg(target_os = "windows")]
+            {
+                let device_id = SYSTEM_AUDIO_DEVICE_ID.to_owned().into();
+
+                if let Some(src) = self.audio_sources.get(&device_id) {
+                    Arc::clone(src)
+                } else {
+                    let src = Arc::new(AudioSource::new(
+                        device_id.clone(),
+                        self.audio_device_module.create_display_audio_source(
+                            &device_id,
+                            &processing,
+                        )?,
+                        processing,
+                    ));
+                    self.audio_sources.insert(device_id, Arc::clone(&src));
+
+                    src
+                }
             }
-
-            self.audio_device_module.recording_device_name(0)?.1.into()
-        };
-
-        let Some(device_index) =
-            self.get_index_of_audio_recording_device(&device_id)?
-        else {
-            bail!(
-                "Cannot find audio device with the specified ID: `{device_id}`",
-            );
-        };
-
-        let src = if let Some(src) = self.audio_sources.get(&device_id) {
-            src.update_audio_processing(&caps.processing);
-            Arc::clone(src)
         } else {
-            let processing =
-                sys::AudioProcessing::new((&caps.processing).into())?;
+            let device_id = if let Some(device_id) = caps.device_id.clone() {
+                device_id.into()
+            } else {
+                // `AudioDeviceModule` is not capturing anything at the moment,
+                // so the first available device (with `0` index) will be used.
+                if self.audio_device_module.recording_devices() < 1 {
+                    bail!("Cannot find any available audio input device");
+                }
 
-            let src = Arc::new(AudioSource::new(
-                device_id.clone(),
-                self.audio_device_module
-                    .create_audio_source(device_index, &processing)?,
-                processing,
-            ));
-            self.audio_sources.insert(device_id, Arc::clone(&src));
+                self.audio_device_module.recording_device_name(0)?.1.into()
+            };
 
-            src
+            let Some(device_index) =
+                self.get_index_of_audio_recording_device(&device_id)?
+            else {
+                bail!(
+                    "Cannot find audio device with the specified ID: \
+                    `{device_id}`",
+                );
+            };
+
+            if let Some(src) = self.audio_sources.get(&device_id) {
+                src.update_audio_processing(&caps.processing);
+                Arc::clone(src)
+            } else {
+                let src = Arc::new(AudioSource::new(
+                    device_id.clone(),
+                    self.audio_device_module
+                        .create_audio_source(device_index, &processing)?,
+                    processing,
+                ));
+                self.audio_sources.insert(device_id, Arc::clone(&src));
+
+                src
+            }
         };
 
         Ok(src)